apps/api/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
Django==6.0.1
djangorestframework==3.16.1
openpyxl==3.1.5
pandas==3.0.1
pillow==11.3.0
requests==2.32.5
reportlab==4.3.2
selenium==4.40.0
sqlparse==0.5.5
tzdata==2025.3
webdriver-manager==4.0.2
